Tear apport GenericSingleResolutionRenderer from MultiResolutionRendererGeneric.
